src/requirements.txt
# Core Dependencies
semantic-kernel==1.3.0
pinecone-client==3.0.3
azure-identity==1.15.0
openai==1.12.0

# Data Processing
langchain==0.1.9
langchain-community==0.0.24
pypdf2==3.0.1
beautifulsoup4==4.12.3
python-docx==1.1.0
markdown==3.5.2
unstructured==0.12.4

# API & Web
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
httpx==0.26.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
aiosqlite==0.19.0

# Utilities
python-dotenv==1.0.1
tenacity==8.2.3
loguru==0.7.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
httpx==0.26.0

# Monitoring
prometheus-client==0.19.0